Refactor trace_utils and module_utils out of tf_test_utils/tf_utils
