+++ integrations/tensorflow/bindings/python/pyiree/tf/support/tf_test_utils_test.py
@@ -14,45 +14,13 @@
 # limitations under the License.
 """Tests for pyiree.tf.support.tf_test_utils."""
 
-import os
-import tempfile
-
-from absl.testing import parameterized
 import numpy as np
 from pyiree.tf.support import tf_test_utils
 from pyiree.tf.support import tf_utils
 import tensorflow as tf
 
 
-class StatefulCountingModule(tf.Module):
-
-  def __init__(self):
-    self.count = tf.Variable([0.])
-
-  @tf.function(input_signature=[])
-  def increment(self):
-    self.count.assign_add(tf.constant([1.]))
-
-  @tf.function(input_signature=[])
-  def get_count(self):
-    return self.count
-
-  @tf.function(input_signature=[tf.TensorSpec([1])])
-  def increment_by(self, value):
-    self.count.assign_add(value)
-
-  @tf.function(input_signature=[tf.TensorSpec([1]), tf.TensorSpec([1])])
-  def increment_by_max(self, a, b):
-    result = tf.maximum(a, b)
-    self.count.assign_add(result)
-    return result
-
-  @tf.function(input_signature=[])
-  def decrement(self):
-    self.count.assign_sub(tf.constant([1.]))
-
-
-class TfFunctionUnittestModule(tf_test_utils.TestModule):
+class TfFunctionUnitTestModule(tf_test_utils.TestModule):
 
   @tf_test_utils.tf_function_unit_test(input_signature=[])
   def no_args(self):
@@ -100,158 +68,7 @@ class TfFunctionUnittestModule(tf_test_utils.TestModule):
     return tf.matmul(a, b)
 
 
-class TestUtilsTests(tf.test.TestCase, parameterized.TestCase):
-
-  @parameterized.named_parameters([
-      {
-          'testcase_name': 'all the same',
-          'array_c': np.array([0, 1, 2]),
-          'array_d': np.array(['0', '1', '2']),
-          'array_e': np.array([0.0, 0.1, 0.2]),
-          'tar_same': True,
-      },
-      {
-          'testcase_name': 'wrong int',
-          'array_c': np.array([1, 1, 2]),
-          'array_d': np.array(['0', '1', '2']),
-          'array_e': np.array([0.0, 0.1, 0.2]),
-          'tar_same': False,
-      },
-      {
-          'testcase_name': 'wrong string',
-          'array_c': np.array([0, 1, 2]),
-          'array_d': np.array(['a', '1', '2']),
-          'array_e': np.array([0.0, 0.1, 0.2]),
-          'tar_same': False,
-      },
-      {
-          'testcase_name': 'wrong float',
-          'array_c': np.array([0, 1, 2]),
-          'array_d': np.array(['0', '1', '2']),
-          'array_e': np.array([1.0, 0.1, 0.2]),
-          'tar_same': False,
-      },
-  ])
-  def test_recursive_check_same(self, array_c, array_d, array_e, tar_same):
-
-    # yapf: disable
-    ref = {
-        'a': 1,
-        'b': [
-            {'c': np.array([0, 1, 2])},
-            {'d': np.array(['0', '1', '2'])},
-            {'e': np.array([0.0, 0.1, 0.2])}
-        ],
-    }
-    tar = {
-        'a': 1,
-        'b': [
-            {'c': array_c},
-            {'d': array_d},
-            {'e': array_e}
-        ],
-    }
-    # yapf: enable
-    same, _ = tf_test_utils.Trace._check_same(ref, tar, rtol=1e-6, atol=1e-6)
-    self.assertEqual(tar_same, same)
-
-  def test_trace_inputs_and_outputs(self):
-
-    def trace_function(module):
-      # No inputs or outputs
-      module.increment()
-      # Only inputs
-      module.increment_by(np.array([81.], dtype=np.float32))
-      # Only outputs
-      module.get_count()
-
-    module = tf_utils.TfCompiledModule.create_from_class(
-        StatefulCountingModule, tf_utils.BackendInfo('tf'))
-    trace = tf_test_utils.Trace(module, trace_function)
-    trace_function(tf_test_utils.TracedModule(module, trace))
-
-    self.assertIsInstance(trace.calls[0].inputs, tuple)
-    self.assertEmpty(trace.calls[0].inputs)
-    self.assertIsInstance(trace.calls[0].outputs, tuple)
-    self.assertEmpty(trace.calls[0].outputs)
-
-    self.assertAllClose(trace.calls[1].inputs[0], [81.])
-    self.assertAllClose(trace.calls[2].outputs[0], [82.])
-
-  def test_nonmatching_methods(self):
-
-    def tf_function(module):
-      module.increment()
-      module.increment()
-
-    def vmla_function(module):
-      module.increment()
-      module.decrement()
-
-    tf_module = tf_utils.TfCompiledModule.create_from_class(
-        StatefulCountingModule, tf_utils.BackendInfo('tf'))
-    tf_trace = tf_test_utils.Trace(tf_module, tf_function)
-    tf_function(tf_test_utils.TracedModule(tf_module, tf_trace))
-
-    vmla_module = tf_utils.IreeCompiledModule.create_from_class(
-        StatefulCountingModule, tf_utils.BackendInfo('iree_vmla'))
-    vmla_trace = tf_test_utils.Trace(vmla_module, vmla_function)
-    vmla_function(tf_test_utils.TracedModule(vmla_module, vmla_trace))
-
-    with self.assertRaises(ValueError):
-      tf_test_utils.Trace.compare_traces(tf_trace, vmla_trace)
-
-  def test_nonmatching_inputs(self):
-
-    def tf_function(module):
-      module.increment_by(np.array([42.], dtype=np.float32))
-
-    def vmla_function(module):
-      module.increment_by(np.array([22.], dtype=np.float32))
-
-    tf_module = tf_utils.TfCompiledModule.create_from_class(
-        StatefulCountingModule, tf_utils.BackendInfo('tf'))
-    tf_trace = tf_test_utils.Trace(tf_module, tf_function)
-    tf_function(tf_test_utils.TracedModule(tf_module, tf_trace))
-
-    vmla_module = tf_utils.IreeCompiledModule.create_from_class(
-        StatefulCountingModule, tf_utils.BackendInfo('iree_vmla'))
-    vmla_trace = tf_test_utils.Trace(vmla_module, vmla_function)
-    vmla_function(tf_test_utils.TracedModule(vmla_module, vmla_trace))
-
-    same, error_messages = tf_test_utils.Trace.compare_traces(
-        tf_trace, vmla_trace)
-    self.assertFalse(same)
-
-  def test_trace_serialize_and_load(self):
-
-    def trace_function(module):
-      module.increment()
-      module.increment_by(np.array([81.], dtype=np.float32))
-      module.increment_by_max(np.array([81], dtype=np.float32),
-                              np.array([92], dtype=np.float32))
-      module.get_count()
-
-    module = tf_utils.IreeCompiledModule.create_from_class(
-        StatefulCountingModule, tf_utils.BackendInfo('iree_vmla'))
-    trace = tf_test_utils.Trace(module, trace_function)
-    trace_function(tf_test_utils.TracedModule(module, trace))
-
-    with tempfile.TemporaryDirectory() as artifacts_dir:
-      trace_function_dir = tf_test_utils._get_trace_dir(artifacts_dir, trace)
-      trace.serialize(trace_function_dir)
-      self.assertTrue(
-          os.path.exists(os.path.join(trace_function_dir, 'metadata.pkl')))
-      loaded_trace = tf_test_utils.Trace.load(trace_function_dir)
-
-      # Check all calls match.
-      self.assertTrue(tf_test_utils.Trace.compare_traces(trace, loaded_trace))
-
-      # Check all other metadata match.
-      self.assertAllEqual(trace.__dict__.keys(), loaded_trace.__dict__.keys())
-      for key in trace.__dict__.keys():
-        if key != 'calls':
-          self.assertEqual(trace.__dict__[key], loaded_trace.__dict__[key])
+class TestUtilsTests(tf.test.TestCase):
 
   def test_tf_function_unittet(self):
 
@@ -260,9 +77,9 @@ class TestUtilsTests(tf.test.TestCase, parameterized.TestCase):
       def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._modules = tf_test_utils.compile_tf_module(
-            TfFunctionUnittestModule)
+            TfFunctionUnitTestModule)
 
-    TfFunctionUnittestTest.generate_unit_tests(TfFunctionUnittestModule)
+    TfFunctionUnittestTest.generate_unit_tests(TfFunctionUnitTestModule)
     test_case = TfFunctionUnittestTest()
     self.assertTrue(hasattr(test_case, 'test_no_args'))
     self.assertTrue(hasattr(test_case, 'test_default_uniform_inputs'))
@@ -270,7 +87,7 @@ class TestUtilsTests(tf.test.TestCase, parameterized.TestCase):
     self.assertTrue(hasattr(test_case, 'test_custom_input_args'))
     self.assertTrue(hasattr(test_case, 'test_high_tolerance'))
 
-    # Will throw an error if 'atol' and 'rtol' are not set.
+    # Will throw an error if 'atol' is not set.
     test_case = TfFunctionUnittestTest()
     test_case.test_high_tolerance()
