+++ integrations/tensorflow/bindings/python/pyiree/tf/support/tf_utils_test.py
@@ -14,125 +14,87 @@
 # limitations under the License.
 """Tests for pyiree.tf.support.tf_utils."""
 
-import os
-import tempfile
-
-from absl import logging
 from absl.testing import parameterized
 import numpy as np
 from pyiree.tf.support import tf_utils
 import tensorflow as tf
 
 
-class ConstantModule(tf.Module):
-
-  @tf.function(input_signature=[])
-  def meaning(self):
-    return tf.constant([42.])
-
-
-class StatefulCountingModule(tf.Module):
-
-  def __init__(self):
-    self.count = tf.Variable([0.])
-
-  @tf.function(input_signature=[])
-  def increment(self):
-    self.count.assign_add(tf.constant([1.]))
-
-  @tf.function(input_signature=[])
-  def get_count(self):
-    return self.count
-
-
-class RandomInitModule(tf.Module):
-
-  def __init__(self):
-    self.value = tf.Variable(tf.random.uniform([1]))
-
-  @tf.function(input_signature=[])
-  def get(self):
-    return self.value
-
-
 class UtilsTests(tf.test.TestCase, parameterized.TestCase):
 
-  def test_artifact_saving(self):
-    backend_info = tf_utils.BackendInfo('iree_vmla')
-    with tempfile.TemporaryDirectory() as artifacts_dir:
-      tf_module = ConstantModule()
-      iree_compiled_module, compiled_path = (
-          tf_utils._incrementally_compile_tf_module(
-              tf_module, backend_info=backend_info,
-              artifacts_dir=artifacts_dir))
+  @parameterized.named_parameters([('int8_to_i8', np.int8, 'i8'),
+                                   ('int32_to_i32', np.int32, 'i32'),
+                                   ('float32_to_f32', np.float32, 'f32'),
+                                   ('float64_to_f64', np.float64, 'f64')])
+  def test_to_mlir_type(self, numpy_type, mlir_type):
+    self.assertEqual(tf_utils.to_mlir_type(numpy_type), mlir_type)
 
-      artifacts_to_check = [
-          'tf_input.mlir',
-          'iree_input.mlir',
-          compiled_path,
-      ]
-      for artifact in artifacts_to_check:
-        artifact_path = os.path.join(artifacts_dir, artifact)
-        logging.info('Checking path: %s', artifact_path)
-        self.assertTrue(os.path.exists(artifact_path))
+  @parameterized.named_parameters([
+      ('single_i32', [np.array([1, 2], dtype=np.int32)], '2xi32=1 2'),
+      ('single_f32', [np.array([1, 2], dtype=np.float32)], '2xf32=1.0 2.0'),
+  ])
+  def test_save_input_values(self, inputs, inputs_str):
+    self.assertEqual(tf_utils.save_input_values(inputs), inputs_str)
+
+  def test_apply_function(self):
+    inputs = [1, [2, 3], (4, 5), {'6': 6, '78': [7, 8]}]
+    expected = [0, [1, 2], (3, 4), {'6': 5, '78': [6, 7]}]
+    result = tf_utils.apply_function(inputs, lambda x: x - 1)
+    self.assertEqual(result, expected)
+    self.assertNotEqual(inputs, expected)
 
   @parameterized.named_parameters([
       {
-          'testcase_name': 'tensorflow',
-          'backend_name': 'tf',
+          'testcase_name': 'all the same',
+          'array_c': np.array([0, 1, 2]),
+          'array_d': np.array(['0', '1', '2']),
+          'array_e': np.array([0.0, 0.1, 0.2]),
+          'tar_same': True,
       },
       {
-          'testcase_name': 'vmla',
-          'backend_name': 'iree_vmla',
+          'testcase_name': 'wrong int',
+          'array_c': np.array([1, 1, 2]),
+          'array_d': np.array(['0', '1', '2']),
+          'array_e': np.array([0.0, 0.1, 0.2]),
+          'tar_same': False,
+      },
+      {
+          'testcase_name': 'wrong string',
+          'array_c': np.array([0, 1, 2]),
+          'array_d': np.array(['a', '1', '2']),
+          'array_e': np.array([0.0, 0.1, 0.2]),
+          'tar_same': False,
+      },
+      {
+          'testcase_name': 'wrong float',
+          'array_c': np.array([0, 1, 2]),
+          'array_d': np.array(['0', '1', '2']),
+          'array_e': np.array([1.0, 0.1, 0.2]),
+          'tar_same': False,
       },
   ])
-  def test_unaltered_state(self, backend_name):
-    backend_info = tf_utils.BackendInfo(backend_name)
-    module = backend_info.compile_from_class(StatefulCountingModule)
+  def test_recursive_check_same(self, array_c, array_d, array_e, tar_same):
 
-    # Test that incrementing works properly.
-    self.assertEqual([0.], module.get_count())
-    module.increment()
-    self.assertEqual([1.], module.get_count())
-
-    module.reinitialize()
-    # Test reinitialization.
-    self.assertEqual([0.], module.get_count())
-
-  def test_to_mlir_type(self):
-    self.assertEqual('i8', tf_utils.to_mlir_type(np.dtype('int8')))
-    self.assertEqual('i32', tf_utils.to_mlir_type(np.dtype('int32')))
-    self.assertEqual('f32', tf_utils.to_mlir_type(np.dtype('float32')))
-    self.assertEqual('f64', tf_utils.to_mlir_type(np.dtype('float64')))
-
-  def test_save_input_values(self):
-    inputs = [np.array([1, 2], dtype=np.int32)]
-    self.assertEqual('2xi32=1 2', tf_utils.save_input_values(inputs))
-    inputs = [np.array([1, 2], dtype=np.float32)]
-    self.assertEqual('2xf32=1.0 2.0', tf_utils.save_input_values(inputs))
-
-  @parameterized.named_parameters([
-      {
-          'testcase_name': 'tensorflow',
-          'backend_name': 'tf',
-      },
-      {
-          'testcase_name': 'vmla',
-          'backend_name': 'iree_vmla',
-      },
-  ])
-  def test_random_initialization(self, backend_name):
-    backend_info = tf_utils.BackendInfo(backend_name)
-
-    # Test compilation is the same.
-    module_1 = backend_info.compile_from_class(RandomInitModule)
-    module_2 = backend_info.compile_from_class(RandomInitModule)
-    self.assertAllEqual(module_1.get(), module_2.get())
-
-    # Test reinitialization is the same.
-    old_value = module_1.get()
-    module_1.reinitialize()
-    self.assertAllEqual(old_value, module_1.get())
+    # yapf: disable
+    ref = {
+        'a': 1,
+        'b': [
+            {'c': np.array([0, 1, 2])},
+            {'d': np.array(['0', '1', '2'])},
+            {'e': np.array([0.0, 0.1, 0.2])}
+        ],
+    }
+    tar = {
+        'a': 1,
+        'b': [
+            {'c': array_c},
+            {'d': array_d},
+            {'e': array_e}
+        ],
+    }
+    # yapf: enable
+    same, _ = tf_utils.check_same(ref, tar, rtol=1e-6, atol=1e-6)
+    self.assertEqual(tar_same, same)
 
 
 if __name__ == '__main__':
